+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/features/narrator_menu/NarratorMenuGUI.java
@@ -1,4 +1,4 @@
-package com.github.khanshoaib3.minecraft_access.features.NarratorMenu;
+package com.github.khanshoaib3.minecraft_access.features.narrator_menu;
 
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.client.gui.widget.ButtonWidget;
