+++ common/src/main/java/com/github/khanshoaib3/minecraft_access/features/point_of_interest/POIBlocks.java
@@ -1,4 +1,4 @@
-package com.github.khanshoaib3.minecraft_access.features.PointOfInterest;
+package com.github.khanshoaib3.minecraft_access.features.point_of_interest;
 
 import com.github.khanshoaib3.minecraft_access.MainClass;
 import com.google.common.collect.ImmutableSet;
